requirements.txt
PyQt5>=5.15.0
Pillow>=8.0.0
reportlab>=3.6.0
pytest-qt>=4.0.0
